requirements.txt
feedparser==6.0.10
pyyaml==6.0.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dateutil==2.8.2
